Fixes and Updated Postman Dev Collection

+++ crates/location_tracking_service/src/common/types.rs
@@ -147,7 +147,7 @@ pub struct AppState {
     pub bucket_expiry: u64,
     pub redis_expiry: u32,
     pub min_location_accuracy: i32,
-    pub last_location_timstamp_expiry: usize,
+    pub last_location_timstamp_expiry: u32,
     pub location_update_limit: usize,
     pub location_update_interval: u64,
     pub producer: Option<FutureProducer>,
